Changed the ifdef for the test program code within fbulib.c to test its own logic from PRG_SAMPLE to PRG_TEST
Added the rule to make the library explicitly now

Moved the code related to printing of fb related info in open_fb into info_fb
so that it can be called seperately if required. However it won't be valid
till open_fb is called, as it just uses variables which are set by open_fb.

Added the fbulib.h into git, which I had forgotten before